v1.1 首发 Pre-release。涵盖 P0（资源包热更 + GDPR 真流式 NDJSON）、P1（领域事件总线）、
P2（重构 / 性能 / 文档 / clippy 清零）+ P0 收尾（流式 GDPR + admin auth 注释）。
版本号 0.6.0-beta.4 -> 1.1.0-beta.1。

+++ tests/amas_effectiveness.rs
@@ -73,7 +73,7 @@ fn test_amas_beats_fixed_interval() {
                 break;
             }
             simulate_review(&mut amas_states[idx], 0.85, now_ms, &config);
-            let interval_ms = compute_interval(&amas_states[idx], 0.85, 1.0, &config) as i64 * 1000;
+            let interval_ms = compute_interval(&amas_states[idx], 0.85, 1.0, &config) * 1000;
             next_review[idx] = now_ms + interval_ms;
             reviewed += 1;
         }
